trpd: use diesel for database interaction
diesel provides multiple advantages:

1. diesel has support for schema migrations.

2. diesel exposes the same API for various database
   backends. This eases support of multiple backends.

3. diesel can be used as a query builder that
   understands the underlying database schema.

4. diesel can auto-generate mappings between database
   schema and rust structs.

For smooth transition `QueryableByName` is used.
https://docs.rs/diesel/1.4.5/diesel/deserialize/trait.QueryableByName.html

It allows to use existing queries as-is via `sql_query()`.
In contrast to `rusqlite` the mapping information is
added to the struct definitions using attribute macros.
The attribute macros can be decreased in verbosity once
diesel is used for schema migrations because then schema
information for rust can be auto-generated.

`QueryableByName` requires all names to be unique to match
query results. This was only not the case for `VaspInfo`
where `BeneficiaryVaspInfo` has been introduced to
work around this problem. Other querying methods in
diesel don't match by name and therefore can be explored
to cleanup that workaround in the future.

Furthermore, the current approach also requires that every
type implements the `FromSql` and `ToSql` traits. That's
why an `ID` wrapper type has been created for `juniper::ID`.
Instead of creating a wrapper type for `url::Url` an
intermediate struct has been used for deserializing `Vasp`.

`SqliteDb::open()` has been fixed in a quick and dirty manner.
In an upcoming change all logic managing the database schema
will be replaced by diesel's schema migration tooling.