21: use diesel master
The `lei` crate is used in both `trpd` and `ovs`. Since
it implements `diesel` traits it forces both crates to
update to master if one of them is updated to master.

`diesel` master is required for `ovs` because dynamic
`bind()` calls will be used in an upcoming change.
This in turn requires `into_boxed()` which is only
available on master.