21: fix clippy::indexing-slicing
Untrusted data should never be accessed using
the index operator because it will panic on
an out-of-bounds access. Instead, `get()` needs
to be used that allows to handle invalid data
gracefully. Luckily, `clippy::indexing-slicing`
can be used to enforce this.

Slice access that is known to never panic is
accessed using `get().unwrap()`. That way the
assumptions are made explicit to the developer.

`first()` is used instead of `get(0)` where
possible.

Tests are exempt from this lint rule by disabling
it on the test module level. When tests have only
a few occurrences then `get().unwrap()` was
preferred because lint exemptions tend to outdate
quickly without a developer noticing.

Also, bump the `rust-crates` version that incorporates
the same lint fixes. This allows us to enable the
lint on the CI.