21: use all juniper crates from crates.io
Finally, `juniper_rocket_async` doesn't depend on a
git dependency and was thus released on crates.io
as `juniper_rocket` replacing the bindings targeting
rocket 0.4.

Also, the bump to tokio 1.x has been merged after a
way long too long time. This allows to drop the dependency
on the custom-patched version of `juniper_graphql_ws`
and used the recently released one from crates.io. As
a consequence the patch sections in `Cargo.toml` can
be removed.