lei: implement FromSql/ToSql without temporary value
Before, the `FromSql`/`ToSql` traits have been implemented
via the `util::sql_via_string!` macro. The macro uses
`to_string()` to crate a temporary string that is then
used to delegate (de-)serialization to existing string
(de-)serialization methods.

Handling temporary values in diesel's `ToSql` trait `impl`
requires specific implementations for Sqlite and Postgres
each. This in turn requires to enable diesel's `postgres`
and `sqlite` features (the `lei` crate fails to build in
isolation as of now) which pulls in the C libraries that
don't compile for wasm targets.

Now, the use of the temporary value is removed by borrowing
the `LEI` types inner string value. This allows to implement
`ToSql` generically for any backend which remedies the tail
of implications described above.