ci: check unused dependencies with cargo-udeps
`cargo-udeps` is useful to detect unused dependencies.
Because it requires nightly features we set the
`RUSTC_BOOTSTRAP` environment variable to unlock those
features on a stable compiler. This is obviously a bit
hacky but the overhead of installing a nightly compiler
just to run an analysis tool is questionable.

Also, remove an unused dependency from the `lei` crate.