21: build diesel without default features
Currently, the default features in diesel are

* `with-deprecated` which we can stop depending on by avoiding
  deprecated import paths.

* `32-column-tables` which is only used by `trpd` and `ovs`.
  Making this explicit is useful to signal which crates are
  depending on it.

The import paths have also been fixed in `sql-enum-macro`
which consequently received a revision bump.

Some crates would still require the `postgres` feature which
is sneaked in by `sql-enum-macro`. Because we plan to make
the macro generic over the database backend we let depended
crates select that feature for us.