21: bump diesel to latest master
Some diesel types have been made private and are now
accessible through different module paths. We also
learned about the `helper_types` module which contains
convenience type aliases that allow to make some
expressions more concise.

Note that on the master branch a regression was introduced
[1] that prematurely disconnected connections in the r2d2
connection pool due to a bug in r2d2. This broke our tests
because we use a r2d2 connection pool of size 1 that connects
to an in-memory Sqlite database. Hence, all data is lost
once the connection pool disconnects from the in-memory database.
Fortunately, this was fixed in [2].

[1] https://github.com/diesel-rs/diesel/pull/2935
[2] https://github.com/diesel-rs/diesel/commit/fb70fd70791010ab9fe8d02c21d78d6946f16dd7