21/graphql: migrate from juniper to async-graphql
juniper has proven to have some rough edges, for example
inscrutable object validation error messages. Unfortunately,
maintenance on such important issues is sluggish, see e.g. [1].

Therefore, this commit migrates the codebase to use the
async-graphql dependency. Due to the schema aggregation under the
autod umbrella, this needs to be done in one go.

Some of the most important differences between juniper and
async-graphql and their impact on the diff are:

 - async-graphql allows using the same struct as both
   input and output type, disambiguated in the schema by
   `name` and `input_name` directives to async-graphql
   declarative macros

 - async-graphql does not allow coercing multiple congruent
   structs into the same graphql type [2]. This introduces
   types like `OfflineAddress`, `EmailAddress` into the
   schema where before a single `Address` could be used

 - async-graphql makes better use of existing `serde`
   (de-)serialization implementations, through the
   `scalar!` macro

 - async-graphql currently has a bug [3] where two operations
   in a multi-operation document cannot use the same names for
   parameters of different types, leading to some necessary
   renames in integration test `queries.graphql` files

We take the opportunity to replace the `graphql-ws-server` glue
layer with usage of axum's websocket support, conveniently exposed
through async-graphql's axum integration. The exposed routes now
need to match the ones coming from either Caddy / Nginx. Before,
the path was completely ignored by the glue layer. Furthermore,
the axum implementation requires [4] the `sec-websocket-protocol`
header, which we now set in the integration test client. The
server implementation will also set the header on the response,
as was done by hand in the glue layer implementation. This ensures
no regression on a Chrome-specific issue [5].

Finally, a nice side-effect for the developer is that async-graphql
ships the recent graphiql 2, whereas juniper had the ancient 0.17.5.

[1] https://github.com/graphql-rust/juniper/issues/693

[2] https://github.com/async-graphql/async-graphql/issues/1015

[3] https://github.com/async-graphql/async-graphql/issues/1014

[4] https://github.com/async-graphql/async-graphql/blob/70049ae798ce015db9a41b637142af990bffaaf0/integrations/axum/src/subscription.rs#L33-L48

[5] https://gitlab.com/21analytics/21-travel-web-ui/-/issues/397