lei/registration authority: point into const array
As mentioned in the code comment, the list of valid
registration authorities must be sorted, which
allows us to look for the index with binary search.
We ensure this with a unit-test.

An alternative that was explored was to have a
`&'static str` field. However, this leads to
problems for deserializer lifetimes, and the
workaround from [1] would need to be applied
to numerous types.

[1] https://github.com/serde-rs/serde/issues/1712